Implement dayofyear and month climatology (#1)
