+++ eexarray/__init__.py
@@ -1,4 +1,5 @@
-from eexarray.Image import Image
-from eexarray.ImageCollection import ImageCollection
+from eexarray.climatology import ClimatologyCollection
+from eexarray.image import Image
+from eexarray.time_series import TimeSeriesCollection
 
 __version__ = "0.0.1"
